update page to remove uneeded reducers and actions
