+++ app/javascript/components/widget/components/widget-header/widget-header.js
@@ -3,15 +3,13 @@ import moment from 'moment';
 
 import shareActions from 'components/share/share-actions';
 import modalMetaActions from 'components/modal-meta/modal-meta-actions';
-import countryRootActions from 'pages/country/root/root-actions';
 import mapActions from 'components/map/map-actions';
 import WidgetHeaderComponent from './widget-header-component';
 
 const actions = {
   ...shareActions,
   ...modalMetaActions,
-  ...mapActions,
-  ...countryRootActions
+  ...mapActions
 };
 
 const mapStateToProps = ({ location, modalMeta }, ownProps) => {
