--- app/javascript/pages/country/page/page-reducers.js
@@ -1,19 +1,0 @@
-export const initialState = {
-  gfwHeaderHeight: 58,
-  showMapMobile: false
-};
-
-const setFixedMapStatus = (state, { payload }) => ({
-  ...state,
-  isMapFixed: payload
-});
-
-const setShowMapMobile = (state, { payload }) => ({
-  ...state,
-  showMapMobile: payload
-});
-
-export default {
-  setFixedMapStatus,
-  setShowMapMobile
-};
